feat- Reorganize the code structure
Reorganize the code structure for testing and maintenance.
