+++ gitcache/e2e/e2e_test.go
@@ -1,0 +1,26 @@
+// Copyright 2020 Security Scorecard Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+package e2e
+
+import (
+	. "github.com/onsi/ginkgo"
+	_ "github.com/onsi/gomega"
+)
+
+var _ = Describe("E2E TEST:gitcache", func() {
+	Context("E2E TEST:Validating gitcache", func() {
+		It("Should be able to connect to the blob store", func() {
+		})
+	})
+})
