+++ gitcache/go.mod
@@ -6,10 +6,15 @@ require (
 	github.com/go-git/go-git/v5 v5.2.0
 	github.com/go-vela/archiver v2.1.0+incompatible // indirect
 	github.com/mholt/archiver v3.1.1+incompatible
+	github.com/nxadm/tail v1.4.8 // indirect
+	github.com/onsi/ginkgo v1.15.0
+	github.com/onsi/gomega v1.10.1
 	github.com/pierrec/lz4 v2.6.0+incompatible // indirect
 	github.com/pkg/errors v0.9.1
 	go.uber.org/zap v1.16.0
 	gocloud.dev v0.22.0
+	golang.org/x/sys v0.0.0-20210303074136-134d130e1a04 // indirect
+	golang.org/x/tools v0.1.0 // indirect
 )
 
 // The originial repository does not handle symlinks and there is an open PR https://github.com/mholt/archiver/pull/230.
