+++ gitcache/pkg/blob.go
@@ -1,4 +1,17 @@
-package main
+// Copyright 2020 Security Scorecard Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+package pkg
 
 import (
 	"context"
@@ -12,7 +25,6 @@ import (
 )
 
 // Cache that use the blob storage like GCS or S3.
-// TODO - Handle returning errors.
 type Cache struct {
 	Bucket *blob.Bucket
 }
