+++ gitcache/pkg/storagepath.go
@@ -1,4 +1,17 @@
-package main
+// Copyright 2020 Security Scorecard Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+package pkg
 
 import (
 	"fmt"
@@ -22,8 +35,8 @@ type StoragePath struct {
 	BlobArchivePath    string // blob path for storing the archive file BlobArchiveFile
 }
 
-// getStoragePath returns path for blob, archiving and also creates temp directories for archiving.
-func getStoragePath(repo RepoURL) (StoragePath, error) {
+// NewStoragePath returns path for blob, archiving and also creates temp directories for archiving.
+func NewStoragePath(repo RepoURL) (StoragePath, error) {
 	cwd, err := os.Getwd()
 	if err != nil {
 		return StoragePath{}, errors.Wrap(err, "unable to the current working dir")
